--- a/.travis.yml
+++ b/.travis.yml
@@ -7,6 +7,4 @@ rust:
   - nightly
 after_success:
   - cargo test --features=lightweight --no-default-features
-  - if [ "$TRAVIS_RUST_VERSION" == "nightly" ]; then
-      cargo bench --features=unstable;
-    fi
+  - ./travis-after-success.sh