--- a/src/cases/case/mod.rs
+++ b/src/cases/case/mod.rs
@@ -25,14 +25,14 @@ pub fn to_case_camel_like(convertable_string: String, camel_options: CamelOption
     let mut last_char: char = camel_options.last_char;
     convertable_string.chars()
         .fold("".to_string(), |mut result, character|
-              if char_is_seperator(&character) {
+              if char_is_seperator(character) {
                   new_word = true;
                   result
               } else if character.is_numeric() {
                   new_word = true;
                   result.push(character);
                   result
-              } else if last_char_lower_current_is_upper_or_new_word(new_word, &last_char, &character) {
+              } else if last_char_lower_current_is_upper_or_new_word(new_word, last_char, character) {
                   new_word = false;
                   result = append_on_new_word(result, first_word, character, &camel_options);
                   first_word = false;
@@ -50,7 +50,7 @@ fn to_snake_like_from_snake_like(convertable_string: String, replace_with: &str,
     let mut last_char: char = ' ';
     convertable_string.chars()
         .fold("".to_string(), |mut result, character|
-              if char_is_seperator(&character) {
+              if char_is_seperator(character) {
                   new_word = true;
                   result.push(replace_with.chars().nth(0).unwrap_or('_'));
                   result
@@ -71,7 +71,7 @@ fn to_snake_like_from_camel_or_class(convertable_string: String,
         .enumerate()
         .fold("".to_string(), |acc, char_with_index|
               if requires_seperator(char_with_index, first_character, &convertable_string) {
-                  snake_like_with_seperator(acc, replace_with, char_with_index.1, case)
+                  snake_like_with_seperator(acc, replace_with, &char_with_index.1, case)
               } else {
                   first_character = false;
                   snake_like_no_seperator(acc, char_with_index.1, case)
@@ -80,10 +80,10 @@ fn to_snake_like_from_camel_or_class(convertable_string: String,
 
 #[inline]
 fn append_on_new_word(mut result: String, first_word: bool, character: char, camel_options: &CamelOptions) -> String {
-    if not_first_word_and_has_seperator(first_word, &camel_options) {
+    if not_first_word_and_has_seperator(first_word, camel_options.has_seperator) {
         result.push(camel_options.injectable_char);
     }
-    if first_word_or_not_inverted(first_word, &camel_options) {
+    if first_word_or_not_inverted(first_word, camel_options.inverted) {
         result.push(character.to_ascii_uppercase());
     } else {
         result.push(character.to_ascii_lowercase());
@@ -91,34 +91,30 @@ fn append_on_new_word(mut result: String, first_word: bool, character: char, cam
     result
 }
 
-#[inline]
-fn not_first_word_and_has_seperator(first_word: bool, camel_options: &CamelOptions) -> bool {
-    !first_word && camel_options.has_seperator
+fn not_first_word_and_has_seperator(first_word: bool, has_seperator: bool) -> bool {
+    has_seperator && !first_word
 }
 
-#[inline]
-fn first_word_or_not_inverted(first_word: bool, camel_options: &CamelOptions) -> bool {
-    !camel_options.inverted || first_word
+fn first_word_or_not_inverted(first_word: bool, inverted: bool) -> bool {
+    !inverted || first_word
 }
 
 
-#[inline]
-fn last_char_lower_current_is_upper_or_new_word(new_word: bool, last_char: &char, character: &char) -> bool{
+fn last_char_lower_current_is_upper_or_new_word(new_word: bool, last_char: char, character: char) -> bool{
     new_word ||
         ((last_char.is_lowercase() && character.is_uppercase()) &&
-         (*last_char != ' '))
+         (last_char != ' '))
 }
 
-#[inline]
-fn char_is_seperator(character: &char) -> bool {
-    *character == '-' || *character == '_' || *character == ' '
+fn char_is_seperator(character: char) -> bool {
+    character == '-' || character == '_' || character == ' '
 }
 
 #[inline]
 fn requires_seperator(char_with_index: (usize, char), first_character: bool, convertable_string: &str) -> bool {
-    char_is_uppercase(char_with_index.1) &&
-        !first_character &&
-        next_or_previous_char_is_lowercase(&convertable_string, char_with_index.0)
+    !first_character &&
+        char_is_uppercase(char_with_index.1) &&
+        next_or_previous_char_is_lowercase(convertable_string, char_with_index.0)
 }
 
 #[inline]
@@ -133,7 +129,7 @@ fn snake_like_no_seperator(mut accumlator: String, current_char: char, case: &st
 }
 
 #[inline]
-fn snake_like_with_seperator(mut accumlator: String, replace_with: &str, current_char: char, case: &str) -> String {
+fn snake_like_with_seperator(mut accumlator: String, replace_with: &str, current_char: &char, case: &str) -> String {
     if case == "lower" {
         accumlator.push(replace_with.chars().nth(0).unwrap_or('_'));
         accumlator.push(current_char.to_ascii_lowercase());
@@ -145,12 +141,11 @@ fn snake_like_with_seperator(mut accumlator: String, replace_with: &str, current
     }
 }
 
-#[inline]
 fn next_or_previous_char_is_lowercase(convertable_string: &str, char_with_index: usize) -> bool {
     convertable_string.chars().nth(char_with_index + 1).unwrap_or('A').is_lowercase() ||
         convertable_string.chars().nth(char_with_index - 1).unwrap_or('A').is_lowercase()
 }
-#[inline]
+
 fn char_is_uppercase(test_char: char) -> bool {
     test_char == test_char.to_ascii_uppercase()
 }
@@ -229,12 +224,12 @@ fn test_next_or_previous_char_is_lowercase_false() {
 
 #[test]
 fn snake_like_with_seperator_lowers() {
-    assert_eq!(snake_like_with_seperator("".to_string(), "^", 'c', "lower"), "^c".to_string())
+    assert_eq!(snake_like_with_seperator("".to_string(), "^", &'c', "lower"), "^c".to_string())
 }
 
 #[test]
 fn snake_like_with_seperator_upper() {
-    assert_eq!(snake_like_with_seperator("".to_string(), "^", 'c', "upper"), "^C".to_string())
+    assert_eq!(snake_like_with_seperator("".to_string(), "^", &'c', "upper"), "^C".to_string())
 }
 
 #[test]
@@ -274,123 +269,75 @@ fn requires_seperator_upper_first_wrap_is_safe_current_lower_next_is_too() {
 
 #[test]
 fn test_char_is_seperator_dash() {
-    assert_eq!(char_is_seperator(&'-'), true)
+    assert_eq!(char_is_seperator('-'), true)
 }
 
 #[test]
 fn test_char_is_seperator_underscore() {
-    assert_eq!(char_is_seperator(&'_'), true)
+    assert_eq!(char_is_seperator('_'), true)
 }
 
 #[test]
 fn test_char_is_seperator_space() {
-    assert_eq!(char_is_seperator(&' '), true)
+    assert_eq!(char_is_seperator(' '), true)
 }
 
 #[test]
 fn test_char_is_seperator_when_not() {
-    assert_eq!(char_is_seperator(&'A'), false)
+    assert_eq!(char_is_seperator('A'), false)
 }
 
 #[test]
 fn test_last_char_lower_current_is_upper_or_new_word_with_new_word() {
-    assert_eq!(last_char_lower_current_is_upper_or_new_word(true, &' ', &'-'), true)
+    assert_eq!(last_char_lower_current_is_upper_or_new_word(true, ' ', '-'), true)
 }
 
 #[test]
 fn test_last_char_lower_current_is_upper_or_new_word_last_char_space() {
-    assert_eq!(last_char_lower_current_is_upper_or_new_word(false, &' ', &'-'), false)
+    assert_eq!(last_char_lower_current_is_upper_or_new_word(false, ' ', '-'), false)
 }
 
 #[test]
 fn test_last_char_lower_current_is_upper_or_new_word_last_char_lower_current_upper() {
-    assert_eq!(last_char_lower_current_is_upper_or_new_word(false, &'a', &'A'), true)
+    assert_eq!(last_char_lower_current_is_upper_or_new_word(false, 'a', 'A'), true)
 }
 
 #[test]
 fn test_last_char_lower_current_is_upper_or_new_word_last_char_upper_current_upper() {
-    assert_eq!(last_char_lower_current_is_upper_or_new_word(false, &'A', &'A'), false)
+    assert_eq!(last_char_lower_current_is_upper_or_new_word(false, 'A', 'A'), false)
 }
 
 #[test]
 fn test_last_char_lower_current_is_upper_or_new_word_last_char_upper_current_lower() {
-    assert_eq!(last_char_lower_current_is_upper_or_new_word(false, &'A', &'a'), false)
+    assert_eq!(last_char_lower_current_is_upper_or_new_word(false, 'A', 'a'), false)
 }
 
 #[test]
 fn test_first_word_or_not_inverted_with_first_word() {
-    let options = CamelOptions {
-        new_word: false,
-        last_char: ' ',
-        first_word: false,
-        injectable_char: ' ',
-        has_seperator: false,
-        inverted: false,
-    };
-    assert_eq!(first_word_or_not_inverted(true, &options), true)
+    assert_eq!(first_word_or_not_inverted(true, false), true)
 }
 
 #[test]
 fn test_first_word_or_not_inverted_not_first_word_not_inverted() {
-    let options = CamelOptions {
-        new_word: false,
-        last_char: ' ',
-        first_word: false,
-        injectable_char: ' ',
-        has_seperator: false,
-        inverted: false,
-    };
-    assert_eq!(first_word_or_not_inverted(false, &options), true)
+    assert_eq!(first_word_or_not_inverted(false, false), true)
 }
 
 #[test]
 fn test_first_word_or_not_inverted_not_first_word_is_inverted() {
-    let options = CamelOptions {
-        new_word: false,
-        last_char: ' ',
-        first_word: false,
-        injectable_char: ' ',
-        has_seperator: false,
-        inverted: true,
-    };
-    assert_eq!(first_word_or_not_inverted(false, &options), false)
+    assert_eq!(first_word_or_not_inverted(false, true), false)
 }
 
 #[test]
 fn test_not_first_word_and_has_seperator_is_first_and_not_seperator() {
-    let options = CamelOptions {
-        new_word: false,
-        last_char: ' ',
-        first_word: false,
-        injectable_char: ' ',
-        has_seperator: false,
-        inverted: false,
-    };
-    assert_eq!(not_first_word_and_has_seperator(true, &options), false)
+    assert_eq!(not_first_word_and_has_seperator(true, false), false)
 }
 
 #[test]
 fn test_not_first_word_and_has_seperator_not_first_and_not_seperator() {
-    let options = CamelOptions {
-        new_word: false,
-        last_char: ' ',
-        first_word: false,
-        injectable_char: ' ',
-        has_seperator: false,
-        inverted: false,
-    };
-    assert_eq!(not_first_word_and_has_seperator(false, &options), false)
+    assert_eq!(not_first_word_and_has_seperator(false, false), false)
 }
 
 #[test]
 fn test_not_first_word_and_has_seperator_not_first_and_has_seperator() {
-    let options = CamelOptions {
-        new_word: false,
-        last_char: ' ',
-        first_word: false,
-        injectable_char: ' ',
-        has_seperator: true,
-        inverted: false,
-    };
-    assert_eq!(not_first_word_and_has_seperator(false, &options), true)
+    assert_eq!(not_first_word_and_has_seperator(false, true), true)
 }