--- a/travis-after-success.sh
+++ b/travis-after-success.sh
@@ -15,5 +15,5 @@ if [ "${TRAVIS_PULL_REQUEST_BRANCH:-$TRAVIS_BRANCH}" != "master" ] && [ "$TRAVIS
     cargo bench --features=unstable > benches-variable && \
     cargo install cargo-benchcmp --force && \
     export PATH=/home/travis/.cargo/bin:$PATH && \
-    cargo benchcmp benches-control benches-variable;
+    cargo benchcmp benches-control benches-variable --threshold 7;
 fi