# What it does:
- Refactors by method extract, the core functionality
- Wraps core functions in tests

# Why it does it:
- The code in case.rs was getting quite complex
- It was also only covered by integration tests

# Benchmarks:

## Master:
```shell
test cases::camelcase::tests::bench_camel0                      ... bench:         141 ns/iter (+/- 39)
test cases::camelcase::tests::bench_camel1                      ... bench:         141 ns/iter (+/- 27)
test cases::camelcase::tests::bench_camel2                      ... bench:         151 ns/iter (+/- 48)
test cases::camelcase::tests::bench_is_camel                    ... bench:         169 ns/iter (+/- 63)
test cases::classcase::tests::bench_class_case                  ... bench:       2,511 ns/iter (+/- 2,079)
test cases::classcase::tests::bench_class_from_snake            ... bench:       2,512 ns/iter (+/- 1,049)
test cases::classcase::tests::bench_is_class                    ... bench:       2,539 ns/iter (+/- 1,003)
test cases::kebabcase::tests::bench_is_kebab                    ... bench:         189 ns/iter (+/- 2)
test cases::kebabcase::tests::bench_kebab                       ... bench:         160 ns/iter (+/- 63)
test cases::kebabcase::tests::bench_kebab_from_snake            ... bench:         257 ns/iter (+/- 126)
test cases::pascalcase::tests::bench_is_pascal                  ... bench:         167 ns/iter (+/- 37)
test cases::pascalcase::tests::bench_pascal0                    ... bench:         145 ns/iter (+/- 75)
test cases::pascalcase::tests::bench_pascal1                    ... bench:         142 ns/iter (+/- 54)
test cases::pascalcase::tests::bench_pascal2                    ... bench:         141 ns/iter (+/- 51)
test cases::screamingsnakecase::tests::bench_is_screaming_snake ... bench:         200 ns/iter (+/- 46)
test cases::screamingsnakecase::tests::bench_screaming_snake    ... bench:         168 ns/iter (+/- 55)
test cases::sentencecase::tests::bench_is_sentence              ... bench:         177 ns/iter (+/- 85)
test cases::sentencecase::tests::bench_sentence                 ... bench:         188 ns/iter (+/- 80)
test cases::sentencecase::tests::bench_sentence_from_snake      ... bench:         146 ns/iter (+/- 74)
test cases::snakecase::tests::bench_is_snake                    ... bench:         189 ns/iter (+/- 69)
test cases::snakecase::tests::bench_snake_from_camel            ... bench:         159 ns/iter (+/- 43)
test cases::snakecase::tests::bench_snake_from_snake            ... bench:         266 ns/iter (+/- 107)
test cases::snakecase::tests::bench_snake_from_title            ... bench:         160 ns/iter (+/- 66)
test cases::tablecase::tests::bench_is_table_case               ... bench:       2,530 ns/iter (+/- 821)
test cases::tablecase::tests::bench_table_case                  ... bench:       2,616 ns/iter (+/- 497)
test cases::titlecase::tests::bench_is_title                    ... bench:         176 ns/iter (+/- 111)
test cases::titlecase::tests::bench_title                       ... bench:         187 ns/iter (+/- 20)
test cases::titlecase::tests::bench_title_from_snake            ... bench:         144 ns/iter (+/- 69)
test cases::traincase::tests::bench_is_train                    ... bench:         175 ns/iter (+/- 74)
test cases::traincase::tests::bench_train                       ... bench:         145 ns/iter (+/- 2)
test cases::traincase::tests::bench_train_from_snake            ... bench:         235 ns/iter (+/- 63)
test tests::benchmark_str_deconstantize                         ... bench:       2,663 ns/iter (+/- 740)
test tests::benchmark_str_demodulize                            ... bench:       2,689 ns/iter (+/- 838)
test tests::benchmark_str_deordinalize                          ... bench:         152 ns/iter (+/- 44)
test tests::benchmark_str_is_camel                              ... bench:         173 ns/iter (+/- 73)
test tests::benchmark_str_is_class                              ... bench:       2,447 ns/iter (+/- 904)
test tests::benchmark_str_is_foreign_key                        ... bench:         235 ns/iter (+/- 82)
test tests::benchmark_str_is_kebab                              ... bench:         191 ns/iter (+/- 114)
test tests::benchmark_str_is_screaming_snake                    ... bench:         204 ns/iter (+/- 29)
test tests::benchmark_str_is_sentence                           ... bench:         177 ns/iter (+/- 63)
test tests::benchmark_str_is_snake                              ... bench:         194 ns/iter (+/- 105)
test tests::benchmark_str_is_table                              ... bench:       2,986 ns/iter (+/- 19)
test tests::benchmark_str_is_title                              ... bench:         177 ns/iter (+/- 2)
test tests::benchmark_str_is_train                              ... bench:         195 ns/iter (+/- 114)
test tests::benchmark_str_ordinalize                            ... bench:         182 ns/iter (+/- 35)
test tests::benchmark_str_pluralize                             ... bench:       2,305 ns/iter (+/- 624)
test tests::benchmark_str_singular                              ... bench:         783 ns/iter (+/- 291)
test tests::benchmark_str_to_camel                              ... bench:         141 ns/iter (+/- 67)
test tests::benchmark_str_to_class                              ... bench:       2,424 ns/iter (+/- 784)
test tests::benchmark_str_to_foreign_key                        ... bench:         465 ns/iter (+/- 89)
test tests::benchmark_str_to_kebab                              ... bench:         159 ns/iter (+/- 31)
test tests::benchmark_str_to_screaming_snake                    ... bench:         169 ns/iter (+/- 75)
test tests::benchmark_str_to_sentence                           ... bench:         142 ns/iter (+/- 24)
test tests::benchmark_str_to_snake                              ... bench:         158 ns/iter (+/- 45)
test tests::benchmark_str_to_table                              ... bench:       2,511 ns/iter (+/- 520)
test tests::benchmark_str_to_title                              ... bench:         143 ns/iter (+/- 40)
test tests::benchmark_str_to_train                              ... bench:         143 ns/iter (+/- 65)
test tests::benchmark_string_deconstantize                      ... bench:       2,687 ns/iter (+/- 860)
test tests::benchmark_string_demodulize                         ... bench:       2,696 ns/iter (+/- 890)
test tests::benchmark_string_deordinalize                       ... bench:         176 ns/iter (+/- 47)
test tests::benchmark_string_is_camel                           ... bench:         196 ns/iter (+/- 65)
test tests::benchmark_string_is_class                           ... bench:       2,473 ns/iter (+/- 155)
test tests::benchmark_string_is_foreign_key                     ... bench:         263 ns/iter (+/- 78)
test tests::benchmark_string_is_kebab                           ... bench:         220 ns/iter (+/- 61)
test tests::benchmark_string_is_screaming_snake                 ... bench:         232 ns/iter (+/- 39)
test tests::benchmark_string_is_sentence                        ... bench:         202 ns/iter (+/- 71)
test tests::benchmark_string_is_snake                           ... bench:         220 ns/iter (+/- 37)
test tests::benchmark_string_is_table                           ... bench:       3,027 ns/iter (+/- 822)
test tests::benchmark_string_is_title                           ... bench:         202 ns/iter (+/- 37)
test tests::benchmark_string_is_train                           ... bench:         201 ns/iter (+/- 55)
test tests::benchmark_string_ordinalize                         ... bench:         206 ns/iter (+/- 79)
test tests::benchmark_string_pluralize                          ... bench:       2,343 ns/iter (+/- 736)
test tests::benchmark_string_singular                           ... bench:         818 ns/iter (+/- 184)
test tests::benchmark_string_to_camel                           ... bench:         166 ns/iter (+/- 81)
test tests::benchmark_string_to_class                           ... bench:       2,448 ns/iter (+/- 751)
test tests::benchmark_string_to_foreign_key                     ... bench:         487 ns/iter (+/- 220)
test tests::benchmark_string_to_kebab                           ... bench:         193 ns/iter (+/- 59)
test tests::benchmark_string_to_screaming_snake                 ... bench:         198 ns/iter (+/- 26)
test tests::benchmark_string_to_sentence                        ... bench:         165 ns/iter (+/- 62)
test tests::benchmark_string_to_snake                           ... bench:         195 ns/iter (+/- 38)
test tests::benchmark_string_to_table                           ... bench:       2,538 ns/iter (+/- 350)
test tests::benchmark_string_to_title                           ... bench:         165 ns/iter (+/- 26)
test tests::benchmark_string_to_train                           ... bench:         165 ns/iter (+/- 31)
```

## This branch
```shell
test cases::camelcase::tests::bench_camel0                      ... bench:         145 ns/iter (+/- 28)
test cases::camelcase::tests::bench_camel1                      ... bench:         151 ns/iter (+/- 84)
test cases::camelcase::tests::bench_camel2                      ... bench:         145 ns/iter (+/- 40)
test cases::camelcase::tests::bench_is_camel                    ... bench:         174 ns/iter (+/- 36)
test cases::classcase::tests::bench_class_case                  ... bench:       2,519 ns/iter (+/- 1,528)
test cases::classcase::tests::bench_class_from_snake            ... bench:       2,635 ns/iter (+/- 1,394)
test cases::classcase::tests::bench_is_class                    ... bench:       2,568 ns/iter (+/- 925)
test cases::kebabcase::tests::bench_is_kebab                    ... bench:         198 ns/iter (+/- 94)
test cases::kebabcase::tests::bench_kebab                       ... bench:         181 ns/iter (+/- 90)
test cases::kebabcase::tests::bench_kebab_from_snake            ... bench:         280 ns/iter (+/- 84)
test cases::pascalcase::tests::bench_is_pascal                  ... bench:         169 ns/iter (+/- 43)
test cases::pascalcase::tests::bench_pascal0                    ... bench:         151 ns/iter (+/- 78)
test cases::pascalcase::tests::bench_pascal1                    ... bench:         153 ns/iter (+/- 87)
test cases::pascalcase::tests::bench_pascal2                    ... bench:         149 ns/iter (+/- 79)
test cases::screamingsnakecase::tests::bench_is_screaming_snake ... bench:         211 ns/iter (+/- 40)
test cases::screamingsnakecase::tests::bench_screaming_snake    ... bench:         179 ns/iter (+/- 70)
test cases::sentencecase::tests::bench_is_sentence              ... bench:         182 ns/iter (+/- 34)
test cases::sentencecase::tests::bench_sentence                 ... bench:         203 ns/iter (+/- 114)
test cases::sentencecase::tests::bench_sentence_from_snake      ... bench:         162 ns/iter (+/- 89)
test cases::snakecase::tests::bench_is_snake                    ... bench:         198 ns/iter (+/- 87)
test cases::snakecase::tests::bench_snake_from_camel            ... bench:         181 ns/iter (+/- 45)
test cases::snakecase::tests::bench_snake_from_snake            ... bench:         284 ns/iter (+/- 77)
test cases::snakecase::tests::bench_snake_from_title            ... bench:         172 ns/iter (+/- 104)
test cases::tablecase::tests::bench_is_table_case               ... bench:       2,500 ns/iter (+/- 544)
test cases::tablecase::tests::bench_table_case                  ... bench:       2,480 ns/iter (+/- 977)
test cases::titlecase::tests::bench_is_title                    ... bench:         182 ns/iter (+/- 7)
test cases::titlecase::tests::bench_title                       ... bench:         207 ns/iter (+/- 121)
test cases::titlecase::tests::bench_title_from_snake            ... bench:         159 ns/iter (+/- 12)
test cases::traincase::tests::bench_is_train                    ... bench:         199 ns/iter (+/- 118)
test cases::traincase::tests::bench_train                       ... bench:         159 ns/iter (+/- 84)
test cases::traincase::tests::bench_train_from_snake            ... bench:         254 ns/iter (+/- 40)
test tests::benchmark_str_deconstantize                         ... bench:       2,681 ns/iter (+/- 1,004)
test tests::benchmark_str_demodulize                            ... bench:       2,978 ns/iter (+/- 2,031)
test tests::benchmark_str_deordinalize                          ... bench:         147 ns/iter (+/- 47)
test tests::benchmark_str_is_camel                              ... bench:         174 ns/iter (+/- 87)
test tests::benchmark_str_is_class                              ... bench:       2,459 ns/iter (+/- 683)
test tests::benchmark_str_is_foreign_key                        ... bench:         238 ns/iter (+/- 95)
test tests::benchmark_str_is_kebab                              ... bench:         214 ns/iter (+/- 67)
test tests::benchmark_str_is_screaming_snake                    ... bench:         217 ns/iter (+/- 35)
test tests::benchmark_str_is_sentence                           ... bench:         182 ns/iter (+/- 4)
test tests::benchmark_str_is_snake                              ... bench:         204 ns/iter (+/- 41)
test tests::benchmark_str_is_table                              ... bench:       2,978 ns/iter (+/- 763)
test tests::benchmark_str_is_title                              ... bench:         182 ns/iter (+/- 34)
test tests::benchmark_str_is_train                              ... bench:         204 ns/iter (+/- 107)
test tests::benchmark_str_ordinalize                            ... bench:         198 ns/iter (+/- 65)
test tests::benchmark_str_pluralize                             ... bench:       2,304 ns/iter (+/- 529)
test tests::benchmark_str_singular                              ... bench:         811 ns/iter (+/- 302)
test tests::benchmark_str_to_camel                              ... bench:         145 ns/iter (+/- 49)
test tests::benchmark_str_to_class                              ... bench:       2,415 ns/iter (+/- 421)
test tests::benchmark_str_to_foreign_key                        ... bench:         475 ns/iter (+/- 126)
test tests::benchmark_str_to_kebab                              ... bench:         181 ns/iter (+/- 47)
test tests::benchmark_str_to_screaming_snake                    ... bench:         212 ns/iter (+/- 97)
test tests::benchmark_str_to_sentence                           ... bench:         166 ns/iter (+/- 87)
test tests::benchmark_str_to_snake                              ... bench:         181 ns/iter (+/- 57)
test tests::benchmark_str_to_table                              ... bench:       2,505 ns/iter (+/- 572)
test tests::benchmark_str_to_title                              ... bench:         155 ns/iter (+/- 69)
test tests::benchmark_str_to_train                              ... bench:         155 ns/iter (+/- 64)
test tests::benchmark_string_deconstantize                      ... bench:       2,707 ns/iter (+/- 30)
test tests::benchmark_string_demodulize                         ... bench:       2,724 ns/iter (+/- 644)
test tests::benchmark_string_deordinalize                       ... bench:         171 ns/iter (+/- 44)
test tests::benchmark_string_is_camel                           ... bench:         204 ns/iter (+/- 71)
test tests::benchmark_string_is_class                           ... bench:       2,472 ns/iter (+/- 637)
test tests::benchmark_string_is_foreign_key                     ... bench:         267 ns/iter (+/- 30)
test tests::benchmark_string_is_kebab                           ... bench:         239 ns/iter (+/- 127)
test tests::benchmark_string_is_screaming_snake                 ... bench:         241 ns/iter (+/- 47)
test tests::benchmark_string_is_sentence                        ... bench:         208 ns/iter (+/- 95)
test tests::benchmark_string_is_snake                           ... bench:         228 ns/iter (+/- 142)
test tests::benchmark_string_is_table                           ... bench:       2,986 ns/iter (+/- 711)
test tests::benchmark_string_is_title                           ... bench:         210 ns/iter (+/- 63)
test tests::benchmark_string_is_train                           ... bench:         219 ns/iter (+/- 124)
test tests::benchmark_string_ordinalize                         ... bench:         215 ns/iter (+/- 3)
test tests::benchmark_string_pluralize                          ... bench:       2,297 ns/iter (+/- 520)
test tests::benchmark_string_singular                           ... bench:         844 ns/iter (+/- 370)
test tests::benchmark_string_to_camel                           ... bench:         169 ns/iter (+/- 42)
test tests::benchmark_string_to_class                           ... bench:       2,436 ns/iter (+/- 519)
test tests::benchmark_string_to_foreign_key                     ... bench:         500 ns/iter (+/- 104)
test tests::benchmark_string_to_kebab                           ... bench:         210 ns/iter (+/- 42)
test tests::benchmark_string_to_screaming_snake                 ... bench:         227 ns/iter (+/- 87)
test tests::benchmark_string_to_sentence                        ... bench:         175 ns/iter (+/- 68)
test tests::benchmark_string_to_snake                           ... bench:         211 ns/iter (+/- 99)
test tests::benchmark_string_to_table                           ... bench:       2,540 ns/iter (+/- 770)
test tests::benchmark_string_to_title                           ... bench:         172 ns/iter (+/- 3)
test tests::benchmark_string_to_train                           ... bench:         190 ns/iter (+/- 117)
```